--- base_decorator.rb
@@ -1,12 +1,0 @@
-require './nameable'
-
-class BaseDecorator < Nameable
-  def initialize(nameable)
-    super()
-    @nameable = nameable
-  end
-
-  def correct_name
-    @nameable.correct_name
-  end
-end
