--- capitalize_decorator.rb
@@ -1,8 +1,0 @@
-require './base_decorator'
-require './person'
-
-class CapitalizeDecorator < BaseDecorator
-  def correct_name
-    @nameable.correct_name.capitalize
-  end
-end
