+++ person.rb
@@ -4,7 +4,7 @@ class Person < Nameable
   attr_reader :id
   attr_accessor :name, :age
 
-  def initialize(age, name: 'unknown', parent_permission: true)
+  def initialize(age, name = 'unknown', parent_permission: true)
     super()
     @id = Random.rand(1..1000)
     @age = age
@@ -28,3 +28,4 @@ class Person < Nameable
     false
   end
 end
+
